feat(modelos): Refactor LogisticaInputInserter to handle Demanda and Oferta inputs
- Updated LogisticaInputInserter to accept LogInpDemanda and LogInpOferta.
- Added new mappings and database operations for LogInpDemanda and its parameters.
- Modified LogEjecucionInputRepository to include methods for fetching LogInpDemanda.
- Adjusted LogEjecucionRepository to accommodate changes in input parameters.
- Enhanced LogEjecucionInputService to provide access to LogInpDemanda.
- Introduced new DTOs for LogInpDemanda and LogInpDemParametro.
- Implemented validators for new DTOs to ensure data integrity.
- Created new domain value objects for LogInpDemanda and LogInpDemParametro.
- Updated persistence models to include LogInpDemanda and LogInpDemParametro.
- Added configuration for CalEjecucion to support new relationships.

+++ Function.Blending.Opt/Application/Features/LogEjecucion/DTOs/Requests/Input/LogInpInfoDto.cs
@@ -14,7 +14,7 @@ public sealed record LogInpInfoDto(
   string Asistente,
   string Supervisora,
   string PaisDestino,
-  int CantidadRuma,
+  decimal CantidadRuma,
   string UnidadMedidaRuma,
   int NumeroMovimientos
 );
